requirements.txt
tensorflow==2.10.0
opencv-python
Pillow
Flask
h5py
matplotlib
scikit-image
numpy==1.23.5